Rearanged dsl part of the project
