+++ lib/Language/Main.hs
@@ -1,0 +1,5 @@
+module Language.Main where
+
+import Text.Parsec
+
+type Parser = Parsec String ()
